requirements.txt
pandas>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
ratelimit>=2.2.1
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
better-profanity>=0.7.0
python-dateutil>=2.8.2
typing-extensions>=4.5.0 